Moved dataset API to the separate module
