+++ dataset/src/main/kotlin/org/jetbrains/kotlinx/dl/dataset/audio/wav/WavFileException.kt
@@ -3,7 +3,7 @@
  * Use of this source code is governed by the Apache 2.0 license that can be found in the LICENSE.txt file.
  */
 
-package org.jetbrains.kotlinx.dl.dataset.sound.wav
+package org.jetbrains.kotlinx.dl.dataset.audio.wav
 
 import java.lang.Exception
