+++ dataset/src/main/kotlin/org/jetbrains/kotlinx/dl/dataset/audio/wav/WavFileFormat.kt
@@ -1,9 +1,9 @@
 /*
- * Copyright 2021 JetBrains s.r.o. and Kotlin Deep Learning project contributors. All Rights Reserved.
+ * Copyright 2020 JetBrains s.r.o. and Kotlin Deep Learning project contributors. All Rights Reserved.
  * Use of this source code is governed by the Apache 2.0 license that can be found in the LICENSE.txt file.
  */
 
-package org.jetbrains.kotlinx.dl.dataset.sound.wav
+package org.jetbrains.kotlinx.dl.dataset.audio.wav
 
 /**
  * Wav file format representing the specification of the WAV file that is saved in its metadata header.
