+++ dataset/src/main/kotlin/org/jetbrains/kotlinx/dl/dataset/preprocessor/ImageShape.kt
