+++ dataset/src/main/kotlin/org/jetbrains/kotlinx/dl/dataset/preprocessor/Sharpen.kt
