+++ dataset/src/main/kotlin/org/jetbrains/kotlinx/dl/dataset/preprocessor/image/InterpolationType.kt
